Revised Pong skeleton, using the toolbox package.
A heavily revised and much simplified version of the skeleton for
Pong.

This version uses the new github.com/gophercoders/toolbox package
to move the nitty gritty of the graphics key and window handling
from the main code.

This results is much simpler code skeleton, which can then be filled
with the game logic.